requirements.txt
requirements.txt
---
paho-mqtt>=1.5.0
python-dateutil>=2.8.1
typing-extensions>=3.7.4